Rearranged, look and feel changes
