+++ client/src/pages/Measures.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { withRouter } from 'react-router-dom'
+import { useHistory } from 'react-router-dom'
 import { makeStyles } from '@material-ui/core/styles';
 import AppBar from '@material-ui/core/AppBar';
 import Toolbar from '@material-ui/core/Toolbar';
@@ -11,7 +11,9 @@ import { AddBox, ArrowDownward } from "@material-ui/icons";
 import MaterialTable from "material-table";
 import { forwardRef } from 'react';
 import { useEffect } from 'react';
+import EditIcon from '@material-ui/icons/Edit';
 import axios from 'axios';
+import { Icon } from '@material-ui/core';
 import Check from '@material-ui/icons/Check';
 import ChevronLeft from '@material-ui/icons/ChevronLeft';
 import ChevronRight from '@material-ui/icons/ChevronRight';
@@ -35,7 +37,7 @@ import Select from '@material-ui/core/Select';
 import Chip from '@material-ui/core/Chip';
 import FaceIcon from '@material-ui/icons/Face';
 import DoneIcon from '@material-ui/icons/Done';
-import { Bar } from 'react-chartjs-2';
+
 
 // simple dialog imports
 
@@ -45,15 +47,34 @@ import DialogContent from '@material-ui/core/DialogContent';
 import DialogContentText from '@material-ui/core/DialogContentText';
 import DialogTitle from '@material-ui/core/DialogTitle';
 import TextField from '@material-ui/core/TextField';
-import { checkServerIdentity } from 'tls';
 
 //simple dialog imports - end
-const drawerWidth = 240;
+const tableIcons = {
+    Add: forwardRef((props, ref) => <AddBox {...props} ref={ref} />),
+    Check: forwardRef((props, ref) => <Check {...props} ref={ref} />),
+    Clear: forwardRef((props, ref) => <Clear {...props} ref={ref} />),
+    Delete: forwardRef((props, ref) => <DeleteOutline {...props} ref={ref} />),
+    DetailPanel: forwardRef((props, ref) => <ChevronRight {...props} ref={ref} />),
+    Edit: forwardRef((props, ref) => <Edit {...props} ref={ref} />),
+    Export: forwardRef((props, ref) => <SaveAlt {...props} ref={ref} />),
+    Filter: forwardRef((props, ref) => <FilterList {...props} ref={ref} />),
+    FirstPage: forwardRef((props, ref) => <FirstPage {...props} ref={ref} />),
+    LastPage: forwardRef((props, ref) => <LastPage {...props} ref={ref} />),
+    NextPage: forwardRef((props, ref) => <ChevronRight {...props} ref={ref} />),
+    PreviousPage: forwardRef((props, ref) => <ChevronLeft {...props} ref={ref} />),
+    ResetSearch: forwardRef((props, ref) => <Clear {...props} ref={ref} />),
+    Search: forwardRef((props, ref) => <Search {...props} ref={ref} />),
+    SortArrow: forwardRef((props, ref) => <ArrowDownward {...props} ref={ref} />),
+    ThirdStateCheck: forwardRef((props, ref) => <Remove {...props} ref={ref} />),
+    ViewColumn: forwardRef((props, ref) => <ViewColumn {...props} ref={ref} />)
+};
+
 
 const useStyles = makeStyles(theme => ({
     root: {
         flexGrow: 1,
-        marginTop: '100px'
+        marginTop: '100px',
+
     },
     menuButton: {
         marginRight: theme.spacing(2),
@@ -74,91 +95,25 @@ const useStyles = makeStyles(theme => ({
     table: {
         minWidth: 650,
     },
-    chart: {
-        marginTop: 100,
-        width: '1000px !important',
-        height: '600px !important',
-    },
 }));
 
-function Settings() {
-    const [initialChartState, setInitialChartState] = React.useState(true);
-    const [date, setDate] = React.useState({ key: Date.now() });
-
-
-    const [cState, setCState] = React.useState({
-        labels: ["2020", "2021", "2022", "2023", "2024", "2025"],
-        datasets: [{
-            label: 'HD0',
-            backgroundColor: "#747474",
-            data: []
-        }, {
-            label: 'HD1',
-            backgroundColor: "#ba962b",
-            data: []
-        }, {
-            label: 'HD2',
-            backgroundColor: "#dfd937",
-            data: []
-        }, {
-            label: 'HD3',
-            backgroundColor: "#02bb71",
-            data: []
-        },
-        {
-            label: 'HD4',
-            backgroundColor: "#518a59",
-            data: []
-        },
-        {
-            label: 'HD5',
-            backgroundColor: "#178b15",
-            data: []
-        }]
-    })
-
-    const [baselineTotalCostColumns, setBaselineTotalCostColumns] = React.useState(
+function Measures() {
+    const history = useHistory();
+    const [columns, setColumns] = React.useState(
         [
-            { title: 'Overall', field: 'overall' },
-            { title: '2020(€)', field: '2020', sorting: true },
-            { title: '2021(€)', field: '2021' },
-            { title: '2022(€)', field: '2022' },
-            { title: '2023(€)', field: '2023' },
-            { title: '2024(€)', field: '2024' },
+            { title: 'Measure ID', field: 'measureId', sorting: true },
+            { title: 'Customer', field: 'customerId.customer' },
+            { title: 'External Measure ID', field: 'externalMeasureId' },
+            { title: 'Measure', field: 'measure' },
+            { title: 'Description', field: 'description' },
+            { title: 'Saving Potential (€)', field: 'potential', type: 'numeric' },
+            { title: 'Duration in Months', field: 'durat    ionInMonth', type: 'numeric' },
+            { title: 'Status', field: 'status' },
+            { title: 'Status Lang', field: 'statusLang' },
         ]);
-    const [baselineColumns, setBaselineColumns] = React.useState(
-        [
-            { title: 'Cost Type', field: 'costType' },
-            {
-                title: 'Sphere of Action',
-                field: 'sphereOfAction',
-                lookup: { '': '', 'infrastructure': 'Infrastructure', 'applications': 'Applications', 'partner': 'Partner', 'staff': 'Staff' },
-            },
-            { title: '2020(€)', field: '2020', sorting: true },
-            { title: '2021(€)', field: '2021' },
-            { title: '2022(€)', field: '2022' },
-            { title: '2023(€)', field: '2023' },
-            { title: '2024(€)', field: '2024' },
-        ]);
-
-    const [savingsColumns, setSavingsColumns] = React.useState(
-        [
-            { title: 'Cost Type', field: 'costType' },
-            {
-                title: 'Sphere of Action',
-                field: 'sphereOfAction',
-                lookup: { '': '', 'infrastructure': 'Infrastructure', 'applications': 'Applications', 'partner': 'Partner', 'staff': 'Staff' },
-            },
-            { title: '2020(€)', field: '2020', sorting: true },
-            { title: '2021(€)', field: '2021' },
-            { title: '2022(€)', field: '2022' },
-            { title: '2023(€)', field: '2023' },
-            { title: '2024(€)', field: '2024' },
-        ])
-
-    const [rows, setRows] = React.useState([]);
-    const [savingsCostTypeRows, setSavingsCostTypeRows] = React.useState([]);
-    const [baselineTotalCostRows, setBaselineTotalCostRows] = React.useState([{ 2020: 123213, 2021: 12353, 2022: 453445, 2023: 123234, 2024: 23234 }]);
+    const [rows, setRows] = React.useState({
+        rows: []
+    });
     const [isAdmin, setAdmin] = React.useState(false);
     const [customers, setCustomers] = React.useState('');
     const [customerOptions, setCustomerOptions] = React.useState([]);
@@ -274,53 +229,35 @@ function Settings() {
         // const result = await axios.get(
         //   'http://localhost:4000/measures',
         // );
-
+        if (!localStorage.getItem('id')) {
+            window.location.href = "/signin"
+            return
+        }
+        if (localStorage.getItem('emailId') && localStorage.getItem('isAdmin') == true) {
+            setAdmin(true)
+        }
+        fetch('http://localhost:4000/measures?id=' + localStorage.getItem('id') + '&customer=' + localStorage.getItem('customerId'), {
+            method: 'GET',
+            headers: { 'x-access-token': localStorage.getItem('token') || '' }
+        })
+            .then(res => res.json())
+            .then(data => {
+                console.log(data)
+                setRows(data);
+            })
+            .catch(error => console.log(error));
 
 
         axios
-            .get("http://localhost:4000/measuredetails" + '?customer=' + localStorage.getItem('customerId'), {
+            .get("http://localhost:4000/customers", {
             })
             .then((response) => {
-                console.log('Response here' + response.data)
-                var year = { '2020': { 'HD0': 0, 'HD1': 0, 'HD2': 0, 'HD3': 0, 'HD4': 0, 'HD5': 0 }, '2021': { 'HD0': 0, 'HD1': 0, 'HD2': 0, 'HD3': 0, 'HD4': 0, 'HD5': 0 }, '2022': { 'HD0': 0, 'HD1': 0, 'HD2': 0, 'HD3': 0, 'HD4': 0, 'HD5': 0 }, '2023': { 'HD0': 0, 'HD1': 0, 'HD2': 0, 'HD3': 0, 'HD4': 0, 'HD5': 0 }, '2024': { 'HD0': 0, 'HD1': 0, 'HD2': 0, 'HD3': 0, 'HD4': 0, 'HD5': 0 } }
-                var hdLevel = ['HD0', 'HD1', 'HD2', 'HD3', 'HD4', 'HD5']
-                var savingsPotential = { 'HD0': 0, 'HD1': 0, 'HD2': 0, 'HD3': 0, 'HD4': 0, 'HD5': 0 }
-                for (var i = 0; i < response.data.length; i++) {
-                    for (var k in year) {
-                        for (var m = 0; m < hdLevel.length; m++) {
-                            if (new Date(response.data[i][hdLevel[m]]).getFullYear() == k) {
-                                savingsPotential[hdLevel[m]] += response.data[i].savingsPotential[0][hdLevel[m]]
-
-                            }
-                        }
-                        for (var h = 0; h < hdLevel.length; h++) {
-                            year[k][hdLevel[h]] += savingsPotential[hdLevel[h]]
-                        }
-
-                        savingsPotential = { 'HD0': 0, 'HD1': 0, 'HD2': 0, 'HD3': 0, 'HD4': 0, 'HD5': 0 }
-                    }
-
-                }
-                console.log(year)
-                var arr = []
-                // for (var y = 0; y < hdLevel.length; year++) {
-                //     for (var i in year) {
-                //         arr.push(year[i][hdLevel[y]])
-                //     }
+                setCustomerOptions(response.data.map((customer) => customer.customer));
+                // var customers = {}
+                // for (var i = 0; i < response.data.length; i++) {
+                //   customers[response.data[i].customer] = []
                 // }
-                for (var j = 0; j < hdLevel.length; j++) {
-                    for (var i in year) {
-                        arr.push(year[i][hdLevel[j]])
-                    }
-                    cState.datasets[j].data = arr
-                    arr = []
-                }
-                console.log(arr)
-                // cState.datasets[0].data = [112, 0, 0, 0, 0]
-                // cState.datasets[1].data = [100, 0, 0, 0, 0]
-                setDate({ key: Date.now() })
-
-                setInitialChartState(false)
+                // setConsultantAccess(customers)
             })
             .catch(function (e) {
                 console.log(e);
@@ -337,23 +274,7 @@ function Settings() {
                 console.log(e);
             });
 
-        axios
-            .get("http://localhost:4000/userrights", {
-            })
-            .then((response) => {
-                // setCustomerOptions(response.data.map((customer) => customer.customer));
-                // var customers = {}
-                // for (var i = 0; i < response.data.length; i++) {
-                //   customers[response.data[i].customer] = []
-                // }
-                setConsultantAccess(response.data)
-                setCustomersAssigned(response.data.filter((consultant) => consultant.email == localStorage.getItem('emailId')))
 
-                setSelectedCustomer(response.data.filter((consultant) => consultant.email == localStorage.getItem('emailId'))[0].customerId)
-            })
-            .catch(function (e) {
-                console.log(e);
-            });
     }, []); const classes = useStyles();
 
 
@@ -460,29 +381,7 @@ function Settings() {
             });
     }
 
-    const handleBaselineColumns = (event) => {
-        setBaselineColumns(columns => [...columns, { title: '2025', field: '2025' },
-        { title: '2026', field: '2026' },
-        { title: '2027', field: '2027' },
-        { title: '2028', field: '2029' },
-        { title: '2029', field: '2029' }]);
-    }
 
-    const handleBaselineTotalCostRows = (event) => {
-        setBaselineTotalCostColumns(columns => [...columns, { title: '2025', field: '2025' },
-        { title: '2026', field: '2026' },
-        { title: '2027', field: '2027' },
-        { title: '2028', field: '2029' },
-        { title: '2029', field: '2029' }]);
-    }
-
-    const handleSavingsColumns = (event) => {
-        setSavingsColumns(columns => [...columns, { title: '2025', field: '2025' },
-        { title: '2026', field: '2026' },
-        { title: '2027', field: '2027' },
-        { title: '2028', field: '2029' },
-        { title: '2029', field: '2029' }]);
-    }
 
     const changeCustomerName = (e) => {
         setCustomerName(e.target.value)
@@ -492,47 +391,80 @@ function Settings() {
         setCustomerDepartment(e.target.value)
     }
     return (
-
         <div className={classes.root}>
+
             <Snackbar open={open} autoHideDuration={6000} onClose={handleClose}>
                 <Alert onClose={handleClose} severity="error">
                     Please enter values for all the fields
         </Alert>
             </Snackbar>
-            {isAdmin == false && localStorage.getItem('isActive') == "true" ?
-                <div className={classes.chart}>
-                    <Bar
-                        redraw
-                        data={cState}
-                        options={{
-                            tooltips: {
-                                displayColors: true,
-                                callbacks: {
-                                    mode: 'x',
-                                },
-                            },
-                            scales: {
-                                xAxes: [{
-                                    barPercentage: 0.7,
-                                    stacked: true,
-                                    gridLines: {
-                                        display: false,
-                                    }
-                                }],
-                                yAxes: [{
-                                    stacked: true,
-                                    ticks: {
-                                        beginAtZero: true,
-                                    },
-                                    type: 'linear',
-                                }]
-                            },
-                            responsive: true,
-                            maintainAspectRatio: false,
-                            legend: { position: 'bottom' },
+            {rows.length > 0 && isAdmin == false && localStorage.getItem('isActive') == "true" ?
+                <div>
+
+                    <Button style={{ marginLeft: "10px", marginTop: "10px", display: "block" }} variant="outlined" color="primary" onClick={() => { history.push('/measuredetails') }}>
+                        Add Measures
+      </Button>
+                    <MaterialTable
+                        icons={tableIcons}
+                        title="Customer Measures"
+                        columns={columns}
+                        actions={[
+                            {
+                                icon: EditIcon,
+                                tooltip: 'Edit',
+                                onClick: (event, rowData) => {
+                                    history.push('/measuredetails/' + rowData.measureId)
+                                }
+                            }
+                        ]}
+                        data={rows}
+                        editable={{
+                            // onRowUpdate: (newData, oldData) =>
+                            //   new Promise(resolve => {
+                            //     history.push('/measuredetails/'+)
+                            // setTimeout(() => {
+                            //   newData.customerId = localStorage.getItem('customerId') + ''
+                            //   resolve();
+                            //   axios
+                            //     .put("http://localhost:4000/measures/" + oldData.measureId, newData, {
+                            //       headers: { 'x-access-token': localStorage.getItem('token') }
+                            //     })
+                            //     .then((response) => {
+                            //       setRows(prevState => {
+                            //         const data = [...prevState];
+                            //         data[data.indexOf(oldData)] = newData;
+                            //         return data;
+                            //       })
+                            //     })
+                            //     .catch(function (e) {
+                            //       console.log(e);
+                            //     }, 600);
+                            // })
+
+                            //}),
+                            onRowDelete: oldData =>
+                                new Promise(resolve => {
+                                    setTimeout(() => {
+                                        resolve();
+                                        axios
+                                            .delete("http://localhost:4000/measuredetails/" + oldData.measureId, {
+                                            })
+                                            .then((response) => {
+                                                setRows(prevState => {
+                                                    const data = [...prevState];
+                                                    data.splice(data.indexOf(oldData), 1);
+                                                    return data;
+                                                })
+                                            })
+                                            .catch(function (e) {
+                                                console.log(e);
+                                            }, 600);
+                                    })
+                                }),
                         }}
                     />
-                </div> : isAdmin == false && localStorage.getItem('isActive') == "true" ? <Typography variant="h6" className={classes.title}>
+                </div>
+                : isAdmin == false && localStorage.getItem('isActive') == "true" ? <Typography variant="h6" className={classes.title}>
                     You have not been assigned customers at the moment, please contact your administrator.
           </Typography>
                     :
@@ -544,4 +476,4 @@ function Settings() {
     );
 }
 
-export default Settings;
+export default Measures;
